# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/izin:
    get:
      operationId: listIzin
      tags: [izin]
      summary: List izin for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of izin
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Izin"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createIzin
      tags: [izin]
      summary: Create an izin under a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IzinInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Izin"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /izin/{id}:
    get:
      operationId: getIzin
      tags: [izin]
      summary: Get an izin with its reports
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Izin detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IzinDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateIzin
      tags: [izin]
      summary: Update an izin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IzinUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Izin"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteIzin
      tags: [izin]
      summary: Delete an izin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    ApprovalStatus:
      type: string
      enum: [draft, submitted, reviewed, approved]
      description: Status alur persetujuan maker-checker-approver
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BusinessScale:
      type: string
      enum: [mikro, kecil, menengah, besar]
      description: Investment scale determining reporting frequency
    OperatingMode:
      type: string
      enum: [pendamping, hibrida, penuh]
      description: Working mode — UMK companion, hybrid, or full human-in-loop
...
    ReportStatus:
      type: string
      enum: [intake, collect, validate, draft, review, submit, monitor, archive]
      description: Pipeline stage of the report
    PeriodType:
      type: string
      enum: [triwulan, semester, tahunan]
...
    RiskLevel:
      type: string
      enum: [rendah, menengah_rendah, menengah_tinggi, tinggi]
      description: Tingkat risiko usaha OSS RBA — rendah, menengah-rendah, menengah-tinggi, tinggi
    BasisPermitType:
      type: string
      enum: [kkpr, persetujuan_lingkungan, pbg, slf, sertifikat_standar, izin]
      description: >-
        Jenis perizinan dasar OSS RBA — KKPR/PKKPR, Persetujuan Lingkungan
        (AMDAL/UKL-UPL/SPPL), PBG, SLF, Sertifikat Standar, atau Izin
    BasisPermitStatus:
      type: string
      enum: [belum_ada, dalam_proses, terbit, kedaluwarsa]
      description: Status kelengkapan perizinan dasar — belum ada, dalam proses, terbit, atau kedaluwarsa
...
    Report:
      type: object
      required:
        - id
        - izinId
        - companyId
        - companyName
        - projectName
        - scale
        - operatingMode
        - periodType
        - periodLabel
        - year
        - deadline
        - status
        - idIzin
        - createdAt
      properties:
        id:
          type: integer
        izinId:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        projectName:
          type: ["string", "null"]
          description: Nama proyek dari Izin terkait
        scale:
          $ref: "#/components/schemas/BusinessScale"
        operatingMode:
          $ref: "#/components/schemas/OperatingMode"
        periodType:
          $ref: "#/components/schemas/PeriodType"
        periodLabel:
          type: string
          description: e.g. "TW II 2026" or "Semester I 2026"
        year:
          type: integer
        deadline:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/ReportStatus"
        idIzin:
          type: string
          description: Id Izin / NIB project identifier
        narrative:
          type: ["string", "null"]
          description: Overall constraint narrative for the report
        ossReceipt:
          type: ["string", "null"]
          description: OSS submission receipt reference
        makerName:
          type: ["string", "null"]
        checkerName:
          type: ["string", "null"]
        approverName:
          type: ["string", "null"]
        makerId:
          type: ["string", "null"]
        checkerId:
          type: ["string", "null"]
        approverId:
          type: ["string", "null"]
        approvalStatus:
          $ref: "#/components/schemas/ApprovalStatus"
        createdAt:
          type: string
          format: date-time
...
    Izin:
      type: object
      required:
        - id
        - companyId
        - idIzin
        - scale
        - createdAt
      properties:
        id:
          type: integer
        companyId:
          type: integer
        idIzin:
          type: string
          description: Id Izin / nomor proyek pada OSS
        kbli:
          type: ["string", "null"]
          description: Kode KBLI proyek Izin ini
        scale:
          $ref: "#/components/schemas/BusinessScale"
        projectName:
          type: ["string", "null"]
        projectLocation:
          type: ["string", "null"]
        riskLevel:
          oneOf:
            - $ref: "#/components/schemas/RiskLevel"
            - type: "null"
          description: Tingkat risiko usaha OSS RBA
        createdAt:
          type: string
          format: date-time
    IzinInput:
      type: object
      required:
        - idIzin
        - scale
      properties:
        idIzin:
          type: string
          minLength: 1
        kbli:
          type: string
        scale:
          $ref: "#/components/schemas/BusinessScale"
        projectName:
          type: string
        projectLocation:
          type: string
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
    IzinUpdate:
      type: object
      properties:
        idIzin:
          type: string
          minLength: 1
        kbli:
          type: string
        scale:
          $ref: "#/components/schemas/BusinessScale"
        projectName:
          type: string
        projectLocation:
          type: string
        riskLevel:
          oneOf:
            - $ref: "#/components/schemas/RiskLevel"
            - type: "null"
    IzinDetail:
      type: object
      required:
        - izin
        - reports
        - basisPermits
      properties:
        izin:
          $ref: "#/components/schemas/Izin"
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        basisPermits:
          type: array
          items:
            $ref: "#/components/schemas/BasisPermit"
    BasisPermit:
      type: object
      required:
        - id
        - izinId
        - type
        - status
        - createdAt
      properties:
        id:
          type: integer
        izinId:
          type: integer
        type:
          $ref: "#/components/schemas/BasisPermitType"
        documentNumber:
          type: ["string", "null"]
          description: Nomor dokumen perizinan
        issuedDate:
          type: ["string", "null"]
          format: date
          description: Tanggal terbit (YYYY-MM-DD)
        validUntil:
          type: ["string", "null"]
          format: date
          description: Tanggal berlaku sampai (YYYY-MM-DD)
        status:
          $ref: "#/components/schemas/BasisPermitStatus"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...