# lib/api-zod/src/generated/types/approvalStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Status alur persetujuan maker-checker-approver
 */
export type ApprovalStatus = typeof ApprovalStatus[keyof typeof ApprovalStatus];


export const ApprovalStatus = {
  draft: 'draft',
  submitted: 'submitted',
  reviewed: 'reviewed',
  approved: 'approved',
} as const;