# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStatus } from './approvalStatus';
import type { BusinessScale } from './businessScale';
import type { OperatingMode } from './operatingMode';
import type { PeriodType } from './periodType';
import type { ReportStatus } from './reportStatus';

export interface Report {
  id: number;
  izinId: number;
  companyId: number;
  companyName: string;
  /**
     * Nama proyek dari Izin terkait
     * @nullable
     */
  projectName: string | null;
  scale: BusinessScale;
  operatingMode: OperatingMode;
  periodType: PeriodType;
  /** e.g. "TW II 2026" or "Semester I 2026" */
  periodLabel: string;
  year: number;
  deadline: Date;
  status: ReportStatus;
  /** Id Izin / NIB project identifier */
  idIzin: string;
  /**
     * Overall constraint narrative for the report
     * @nullable
     */
  narrative?: string | null;
  /**
     * OSS submission receipt reference
     * @nullable
     */
  ossReceipt?: string | null;
  /** @nullable */
  makerName?: string | null;
  /** @nullable */
  checkerName?: string | null;
  /** @nullable */
  approverName?: string | null;
  /** @nullable */
  makerId?: string | null;
  /** @nullable */
  checkerId?: string | null;
  /** @nullable */
  approverId?: string | null;
  approvalStatus?: ApprovalStatus;
  createdAt: Date;
}